Failing test for construct new object references
