+++ lib/Indexer/Tests/Adapter/ReferenceFinder/IndexedReferenceFinderTest.php
@@ -233,6 +233,21 @@ class IndexedReferenceFinderTest extends IntegrationTestCase
         ,
             3
         ];
+
+        yield 'show new object expressions when finding references on __construct' => [
+            <<<'EOT'
+                // File: project/foobar.php
+                <?php Foobar::bar();
+
+                // File: project/subject.php
+                <?php class Foobar { public function __c<>onstruct() {} public static function bar() {} }
+
+                // File: project/class1.php
+                <?php new Foobar();
+                EOT
+        ,
+            1
+        ];
     }
 
     /**
